feat: support update beta package version v9
