+++ scripts/update-beta-package-version.js
@@ -1,96 +1,19 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
+const { WorkSpaces } = require("./libs/work-spaces");
+const { cwd } = require("process");
+const dayjs = require("dayjs");
 const { join } = require("path");
-const { existsSync, readdirSync } = require("fs-extra");
-const { readJsonSync } = require("fs-extra");
-const { outputFileSync } = require("fs-extra");
+const { readJSONSync } = require("fs-extra");
 
-class WorkSpaces {
-  /**
-   * @type {string}
-   */
-  rootDir;
+const currentVersion = readJSONSync(
+  join(cwd(), "package.json")
+).version.replace(/-.*/, "");
+const time = dayjs().format("YYYY-MM-DD-HH-MM-SSS");
 
-  /**
-   * @type {string[]}
-   * @example ['core', 'services', 'packages']
-   */
-  workSpaces;
-
-  includeWorkspaceRoot;
-
-  #packageJson = "package.json";
-
-  constructor({ rootDir, workSpaces, includeWorkspaceRoot }) {
-    this.rootDir = rootDir;
-    this.workSpaces = workSpaces;
-    this.includeWorkspaceRoot = includeWorkspaceRoot;
-  }
-
-  async setVersion(version) {
-    return this.#setVersion(version);
-  }
-
-  async #setVersion(version) {
-    const packageJsonPaths = await this.#getPackageJsonPaths();
-
-    await Promise.all(
-      packageJsonPaths.map((packageJsonPath) =>
-        this.#setPackageVersion(packageJsonPath, version)
-      )
-    );
-  }
-
-  /**
-   *
-   * @param {string} packageJsonPath
-   * @param {string} version
-   * @returns {Promise<void>}
-   */
-  async #setPackageVersion(packageJsonPath, version) {
-    const packageJson = readJsonSync(packageJsonPath);
-    packageJson.version = version;
-    outputFileSync(packageJsonPath, JSON.stringify(packageJson, null, 2));
-  }
-
-  /**
-   *
-   * @returns {Promise<string[]>}
-   */
-  async #getPackageJsonPaths() {
-    const packageJsonPaths = [];
-
-    if (
-      this.includeWorkspaceRoot &&
-      existsSync(join(this.rootDir, this.#packageJson))
-    ) {
-      packageJsonPaths.push(join(this.rootDir, this.#packageJson));
-    }
-
-    for (const workspace of this.workSpaces) {
-      if (!existsSync(join(this.rootDir, workspace))) {
-        throw new Error(`workspace(${workspace}) not found`);
-      }
-
-      const files = readdirSync(workspace, { withFileTypes: true });
-
-      for (const file of files) {
-        if (
-          file.isDirectory() &&
-          existsSync(
-            join(this.rootDir, workspace, file.name, this.#packageJson)
-          )
-        ) {
-          packageJsonPaths.push(
-            join(this.rootDir, workspace, file.name, this.#packageJson)
-          );
-        }
-      }
-    }
-
-    return packageJsonPaths;
-  }
-}
-
-module.exports = {
-  WorkSpaces,
-};
+const newVersion = `${currentVersion}-beta-${time}`;
+const workspace = new WorkSpaces({
+  rootDir: cwd(),
+  workSpaces: ["core"],
+  includeWorkspaceRoot: true,
+});
+workspace.setVersion(newVersion);
